+++ docs/js/metro.js
@@ -6754,9 +6754,11 @@ $.widget("metro.slider", {
 
     options: {
         position: 0,
+        buffer: 0,
         accuracy: 0,
         color: 'default',
         completeColor: 'default',
+        bufferColor: 'default',
         markerColor: 'default',
         colors: false,
         showHint: false,
@@ -6765,7 +6767,7 @@ $.widget("metro.slider", {
         vertical: false,
         min: 0,
         max: 100,
-        animate: true,
+        animate: false,
         minValue: 0,
         maxValue: 100,
         currValue: 0,
@@ -6775,6 +6777,7 @@ $.widget("metro.slider", {
         onStartChange: function(){},
         onChange: function(value, slider){},
         onChanged: function(value, slider){},
+        onBufferChange: function(value, slider){},
 
         _slider : {
             vertical: false,
@@ -6805,12 +6808,16 @@ $.widget("metro.slider", {
             }
         });
 
+        element.data('internal_id', uniqueId());
+        //console.log(element.data('internal_id'));
+
         o.accuracy = o.accuracy < 0 ? 0 : o.accuracy;
         o.min = o.min < 0 ? 0 : o.min;
         o.min = o.min > o.max ? o.max : o.min;
         o.max = o.max > 100 ? 100 : o.max;
         o.max = o.max < o.min ? o.min : o.max;
         o.position = this._correctValue(element.data('position') > o.min ? (element.data('position') > o.max ? o.max : element.data('position')) : o.min);
+        o.buffer = this._correctValue(element.data('buffer') > o.min ? (element.data('buffer') > o.max ? o.max : element.data('buffer')) : o.min);
         o.colors = o.colors ? o.colors.split(",") : false;
 
         s.vertical = o.vertical;
@@ -6832,6 +6839,7 @@ $.widget("metro.slider", {
         this._createSlider();
         this._initPoints();
         this._placeMarker(o.position);
+        this._showBuffer(o.buffer);
 
         var event_down = isTouchDevice() ? 'touchstart' : 'mousedown';
 
@@ -6862,16 +6870,6 @@ $.widget("metro.slider", {
         element.on(event_down, function (e) {
             e.preventDefault();
             that._startMoveMarker(e);
-            //if (typeof o.onStartChange === 'function') {
-            //    o.onStartChange();
-            //} else {
-            //    if (typeof window[o.onStartChange] === 'function') {
-            //        window[o.onStartChange]();
-            //    } else {
-            //        var result = eval("(function(){"+o.onStartChange+"})");
-            //        result.call();
-            //    }
-            //}
         });
 
         element.data('slider', this);
@@ -6884,15 +6882,15 @@ $.widget("metro.slider", {
         var event_move = isTouchDevice() ? 'touchmove' : 'mousemove';
         var event_up = isTouchDevice() ? 'touchend' : 'mouseup mouseleave';
 
-        $(document).on(event_move, function (event) {
+        $(element).on(event_move, function (event) {
             that._movingMarker(event);
             if (!element.hasClass('permanent-hint')) {
                 hint.css('display', 'block');
             }
         });
-        $(document).on(event_up, function () {
-            $(document).off('mousemove');
-            $(document).off('mouseup');
+        $(element).on(event_up, function () {
+            $(element).off('mousemove');
+            $(element).off('mouseup');
             element.data('value', o.position);
             element.trigger('changed', o.position);
             element.trigger('change', o.position);
@@ -7113,11 +7111,12 @@ $.widget("metro.slider", {
     _createSlider: function(){
         var element = this.element,
             o = this.options,
-            complete, marker, hint;
+            complete, marker, hint, buffer;
 
         element.html('');
 
         complete = $("<div/>").addClass("complete").appendTo(element);
+        buffer = $("<div/>").addClass("buffer").appendTo(element);
         marker = $("<a/>").addClass("marker").appendTo(element);
 
         if (o.showHint) {
@@ -7136,6 +7135,13 @@ $.widget("metro.slider", {
                 complete.css('background-color', o.completeColor);
             } else {
                 complete.addClass(o.completeColor);
+            }
+        }
+        if (o.bufferColor !== 'default') {
+            if (o.bufferColor.isColor()) {
+                buffer.css('background-color', o.bufferColor);
+            } else {
+                buffer.addClass(o.bufferColor);
             }
         }
         if (o.markerColor !== 'default') {
@@ -7180,9 +7186,66 @@ $.widget("metro.slider", {
                 }
             }
 
+            //if (typeof o.onChanged === 'function') {
+            //    o.onChanged(returnedValue, element);
+            //} else {
+            //    if (typeof window[o.onChanged] === 'function') {
+            //        window[o.onChanged](returnedValue, element);
+            //    } else {
+            //        var result = eval("(function(){"+o.onChanged+"})");
+            //        result.call(returnedValue, element);
+            //    }
+            //}
+
             return this;
         } else {
             returnedValue = o.returnType === 'value' ? this._valueToRealValue(o.position) : o.position;
+            return returnedValue;
+        }
+    },
+
+    _showBuffer: function(value){
+        var size, oldSize, o = this.options, element = this.element,
+            buffer = this.element.children('.buffer');
+
+        oldSize = o.buffer;
+        size = value == 100 ? 99.9 : value;
+
+        if (o._slider.vertical) {
+            this._animate(buffer.css('height', oldSize+'%'),{height: size+'%'});
+
+        } else {
+            this._animate(buffer.css('width', oldSize+'%'),{width: size+'%'});
+        }
+    },
+
+    buffer: function (value) {
+        var element = this.element, o = this.options, returnedValue;
+
+        if (typeof value !== 'undefined') {
+
+            value = value > 100 ? 100 : value;
+            value = value < 0 ? 0 : value;
+
+            this._showBuffer(parseInt(value));
+            o.buffer = parseInt(value);
+
+            returnedValue = o.buffer;
+
+            if (typeof o.onBufferChange === 'function') {
+                o.onBufferChange(returnedValue, element);
+            } else {
+                if (typeof window[o.onBufferChange] === 'function') {
+                    window[o.onBufferChange](returnedValue, element);
+                } else {
+                    var result = eval("(function(){"+o.onBufferChange+"})");
+                    result.call(returnedValue, element);
+                }
+            }
+
+            return this;
+        } else {
+            returnedValue = o.buffer;
             return returnedValue;
         }
     },
@@ -8891,6 +8954,13 @@ $.widget( "metro.video" , {
         video.on("canplay", function(){
             controls.fadeIn();
             preloader.hide();
+            var buffered = video_obj.buffered.length ? Math.round(Math.floor(video_obj.buffered.end(0)) / Math.floor(video_obj.duration) * 100) : 0;
+            that._setBufferSize(buffered);
+        });
+
+        video.on('progress', function(){
+            var buffered = video_obj.buffered.length ? Math.round(Math.floor(video_obj.buffered.end(0)) / Math.floor(video_obj.duration) * 100) : 0;
+            that._setBufferSize(buffered);
         });
 
         video.on("timeupdate", function(){
@@ -8909,6 +8979,7 @@ $.widget( "metro.video" , {
         video.on('ended', function(){
             that._stopVideo();
         });
+
     },
 
     _timeToString: function(time){
@@ -8935,6 +9006,13 @@ $.widget( "metro.video" , {
         var video = element.find("video"), video_obj = video[0];
         var slider = element.find(".stream-slider").data("slider");
         slider.value(Math.round(video_obj.currentTime * 100 / element.data('duration')));
+    },
+
+    _setBufferSize: function(value){
+        var that = this, element = this.element, element_obj = element[0], o = this.options;
+        var video = element.find("video"), video_obj = video[0];
+        var slider = element.find(".stream-slider").data("slider");
+        slider.buffer(Math.round(value));
     },
 
     _stopVideo: function(){
@@ -8995,13 +9073,13 @@ $.widget( "metro.video" , {
             showHint: true,
             animate: false,
             markerColor: 'bg-red',
-            completeColor: 'ribbed-cyan',
+            completeColor: 'bg-cyan',
             onStartChange: function(){
                 video_obj.pause();
             },
             onChanged: function(value, slider){
-                //if (video_obj.seekable.length > 0)
-                video_obj.currentTime = (element.data('duration') * value / 100).toFixed(0);
+                if (video_obj.seekable.length > 0)
+                    video_obj.currentTime = (element.data('duration') * value / 100).toFixed(0);
 
                 if (element.data('played') && video_obj.currentTime > 0)
                     video_obj.play();
@@ -9084,7 +9162,7 @@ $.widget( "metro.video" , {
             showHint: true,
             animate: false,
             markerColor: 'bg-red',
-            completeColor: 'ribbed-green',
+            completeColor: 'bg-green',
             onChange: function(value, slider){
                 video_obj.volume = value/100;
                 that._setupVolumeButton();